+++ modules/search-bar-component/src/search-bar.module.ts
@@ -1,3 +1,4 @@
+import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
 import { SearchBarComponent }   from './search-bar.component';
@@ -8,8 +9,8 @@ import {ProductServiceModule} from '@nodeart/productservice/index';
  * Search bar for searching products by name
  */
 @NgModule({
-    imports: [ProductServiceModule],
-    exports: [],
+    imports: [ProductServiceModule, CommonModule],
+    exports: [SearchBarComponent],
     declarations: [SearchBarComponent],
     providers: [],
 })